finished Lesson11Logic
