+++ Assets/Code/Scripts/SceneScripts/Scene_11/SolarPanel.cs
@@ -6,7 +6,6 @@ using UnityEngine;
 
 public class SolarPanel : MonoBehaviour
 {
-    [SerializeField]
     protected bool _isIntact = true;
 
     public bool IsIntact()
